Refactor: Move CorruptedDatabaseFilesHandler
