+++ common/src/main/java/bisq/common/file/CorruptedDatabaseFilesHandler.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.common.storage;
+package bisq.common.file;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
