+++ core/src/main/java/bisq/core/app/BisqHeadlessApp.java
@@ -20,8 +20,8 @@ package bisq.core.app;
 import bisq.core.trade.TradeManager;
 
 import bisq.common.UserThread;
+import bisq.common.file.CorruptedDatabaseFilesHandler;
 import bisq.common.setup.GracefulShutDownHandler;
-import bisq.common.storage.CorruptedDatabaseFilesHandler;
 
 import com.google.inject.Injector;
