+++ desktop/src/main/java/bisq/desktop/main/MainViewModel.java
@@ -68,7 +68,7 @@ import bisq.common.Timer;
 import bisq.common.UserThread;
 import bisq.common.app.DevEnv;
 import bisq.common.config.Config;
-import bisq.common.storage.CorruptedDatabaseFilesHandler;
+import bisq.common.file.CorruptedDatabaseFilesHandler;
 
 import com.google.inject.Inject;
