+++ p2p/src/test/java/bisq/network/p2p/MockNode.java
@@ -17,8 +17,6 @@
 
 package bisq.network.p2p;
 
-import bisq.network.p2p.seed.SeedNodeRepository;
-
 import bisq.network.p2p.network.Connection;
 import bisq.network.p2p.network.InboundConnection;
 import bisq.network.p2p.network.NetworkNode;
@@ -26,10 +24,11 @@ import bisq.network.p2p.network.OutboundConnection;
 import bisq.network.p2p.network.Statistic;
 import bisq.network.p2p.peers.PeerManager;
 import bisq.network.p2p.peers.peerexchange.PeerList;
+import bisq.network.p2p.seed.SeedNodeRepository;
 
 import bisq.common.ClockWatcher;
+import bisq.common.file.CorruptedDatabaseFilesHandler;
 import bisq.common.proto.persistable.PersistenceProtoResolver;
-import bisq.common.storage.CorruptedDatabaseFilesHandler;
 import bisq.common.storage.Storage;
 
 import java.nio.file.Files;
